docker-compose/auth-nfc-v2/requirements.txt
Flask
Flask-CORS
PyJWT
psycopg2-binary
flask-bcrypt
